+++ index.html
@@ -17,12 +17,20 @@
   <body class="dia">
     <header id="logo">
       <img src="Recursos/logo.svg" alt="logo de el tiempo en 8 bits" />
+      <form>
+        <input
+          type="search"
+          name="location"
+          id="location"
+          placeholder="Indica tu ciudad"
+        />
+      </form>
     </header>
     <main>
       <section class="activo" id="gps">
         <img src="Recursos/GPS-animacion.gif" alt="grafico de pantalla gps" />
         <p>¿Nos dejas ver tu ubicación?</p>
-        <button type="submit" id="gps-boton"><p>PERMITIR GPS</p></button>
+        <button type="submit" id="gps-boton">PERMITIR GPS</button>
       </section>
       <section class="clima"></section>
       <section class="carrusel" id="carrusel"></section>
